backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
redis
pandas
numpy
scikit-learn
yfinance
requests
beautifulsoup4
textblob
python-multipart
python-dotenv
pydantic
pydantic-settings
httpx
websockets
asyncpg
python-dateutil
matplotlib
plotly
nltk
vaderSentiment
websocket-client
APScheduler